### 2016-05-19 #45 @masalmon
monkeylearn: natural language processing with the monkeylearn API
* 1. What does this package do? (explain in 50 words or less)
This package provides access to the existing modules of the monkeylearn API, allowing for instance Named Entity Recognition, language detection, attribution of topics to texts.
* 2. Paste the full DESCRIPTION file inside a code block below.

```
Package: monkeylearn
Title: Access to the Monkeylearn API for text classifiers and extractors
Version: 0.1.0
Authors@R: person("Maëlle", "Salmon", email = "maelle.salmon@yahoo.se", role = c("aut", "cre"))
Description: The package allows using some services of Monkeylearn which is
    a Machine Learning platform on the cloud that allows software companies and
    developers to easily extract actionable data from text.
Suggests: testthat,
Imports: jsonlite, dplyr, httr
```

* 3. URL for the package (the development repository, not a stylized html page)
https://github.com/masalmon/monkeylearn
* 4. What data source(s) does it work with (if applicable)?
It works with monkeylearn API
http://monkeylearn.com/
Note that one can register using one's Github account.
* 5. Who is the target audience?
Anyone that wants to use Natural Language Processing methods without e.g. first installing Java, or someone that already uses monkeylearn and wants to integrate it in a R workflow (but only for text processing, not for module development)
* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?
Yes the openNLP package provides several Natural Language Processing methods such as Named Entity Recognition but has a Java dependency that makes its installation difficult for some users.
* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
    * [ x] This package does not violate the Terms of Service of any service it interacts with.
    * [ x] The repository has continuous integration with Travis CI and/or another service
    * [ x] The package contains a vignette
    * [x ] The package contains a reasonably complete README with `devtools` install instructions
    * [ x] The package contains unit tests
    * [ x] The package only exports functions to the NAMESPACE that are intended for end users
* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
    * [ ] Are there any package dependencies not on CRAN?
    * [ x] Do you intend for this package to go on CRAN?
    * [ x] Does the package have a CRAN accepted license?
    * [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below.
* 9. Please add explanations below for any exceptions to the above:
It's not an explanation to any exception but I wanted to underline I'm not sure I've chosen the best solution for dealing with the throttle limit (I retry 5 times if there is a 429 API error which is throttle limit), so I'd very much appreciate feedback on this (and on anything else, obviously!).
* 10. If this is a resubmission following rejection, please explain the change in circumstances.
