### 2016-05-24 #45 @noamross
Hi @masalmon, I'll be your handling editor and am currently seeking reviewers.  FYI, this submission prompted discussion among the editors on two topics:

-  **Fit**: Previous API wrappers we've accepted had fit under accessing and depositing in databases, geospatial processing, and plotting services. This package fit none of those, but clearly fit in with a growing theme  - extracting data from unstructured sources (`pdftools` and `tabulizer` being examples.).  We've added a new category to our [policies document](https://github.com/ropensci/onboarding/blob/master/policies.md) to explicitly include packages, API wrappers or not, in this category, without expanding to "any API wrapper".
- **API Stability**: With so many processing-as-a-service startups in the market, its inevitable that APIs will go through periods of instability and maturation, and many may be short-lived, and thus package maintenance requirements may be high or packages may become obsolete.  We came to no conclusion from this, but will be keeping an eye on API wrapper packages to see how our system copes with this churn.
