### 2016-05-24 #45 @masalmon
Cool, thanks. Actually [`geoparser`](https://github.com/ropensci/onboarding/issues/43) also extracts data from unstructured text & depends on a processing-as-a-service startup although it's a geospatial thing too.

I know I'm creating work for myself later with API packages. `ropenaq` also breaks when the API changes... I hope I'll be able to cope & that I'll be able to find replacements for the APIs which will disappear :smile_cat:
