### 2016-07-11 #45 @noamross
Additional reviewer: @leeper
