### 2016-07-12 #45 @leeper
This is a slim and easy-to-use package that works exactly as described. The documentation is generally complete and all examples in docs and README run as expected. I think it's really close to being ready for acceptance, but there are a couple of issues and I have some suggestions on usability.

1. I get an error when building vignettes from GitHub (`build_vignettes = TRUE`):

     ```R
     Quitting from lines 35-41 (monkeylearn_intro.Rmd)
     Error: processing vignette 'monkeylearn_intro.Rmd' failed with diagnostics:
     HTTP failure: 401
     Invalid token header. No credentials provided.
     Execution halted
     ```

     This seems to be due to requiring an API key for examples to run. For CRAN release, the code may need to be pre-executed so the vignettes can build. This issue also emerges on `R CMD check` in a few places:

      1. examples, which should probably be wrapped in `\dontrun{}`
      2. test suite, which probably needs to be run conditionally using whichever CRAN-identifying method is appropriate (version number, testthat's check, a simple `if()` based on API key availability, etc.).

2. I also get a NAMESPACE note in `R CMD check`:

     ```
     * checking R code for possible problems ... NOTE
     monkeylearn_check: no visible binding for global variable 'text'
     monkeylearn_parse: no visible global function definition for 'is'
       is text
     Consider adding
       importFrom("graphics", "text")
       importFrom("methods", "is")
     to your NAMESPACE file (and ensure that your DESCRIPTION Imports field
     contains 'methods').
     ```

    That is all pretty minor.

3. The code itself also looks good and is very straightforward, seeing as there are only two functions: `monkeylearn_extract()` and `monkeylearn_classify()`. For both functions, the output is a list of two tbl_dfs. Only the first one `"results"` is really useful. The second, `"headers"`, is basically request metadata. I think it would be more useful to return a single tbl_df and leave the `"headers"` information as attributes to that, in the form of a list. This will improve efficiency because the headers won't be [run through both `as.data.frame()` and `tbl_df()`](https://github.com/masalmon/monkeylearn/blob/master/R/utils.R#L106-L109) and will make the results easier to access.

4. The only purpose for the `headers` response seems to be to track the `x.query.limit.remaining` HTTP header. I would handle this in two ways, both different from current behavior:

 1. Report this value when `verbose = TRUE`.
 2. Store the value in a non-exported package environment and have internal functions `monkeylearn_get_extractor()` and `monkeylearn_get_classify()` update the value after each request and then issue a warning when these values get low.

5. The functions should use `message()` rather than `print()` for diagnostics when `verbose = TRUE`. This occurs in several places:

  - https://github.com/masalmon/monkeylearn/blob/master/R/extractor.R#L64
  - https://github.com/masalmon/monkeylearn/blob/master/R/extractor.R#L75
  - https://github.com/masalmon/monkeylearn/blob/master/R/classify.R#L41
  - https://github.com/masalmon/monkeylearn/blob/master/R/classify.R#L52

6. The only major substantive suggestion I have for the package is to perhaps create some functions or variables to represent the classifier ID and extractor ID values for some common methods. For example, the README and examples require the user to input something like `"ex_isnnZRbS"` as the value of `extractor_id` ala `monkeylearn_extract(request = text, extractor_id = "ex_y7BPYzNG")`. If this could instead be something like `monkeylearn_extract(request = text, extractor_id = .entity)`, `monkeylearn_classify(request = text, classifier_id = .profanity)`, etc. it would greatly simplify the use of the most commonly needed methods.

7. One really small code style comment: on `if ()`, `while ()`, etc. statements, I think Hadley, Google, and most style guides recommend spacing between brackets, so that:

     ```R
     while(class(output) == "try-error" && try_number < 6){
     ...
     }
     ```

     becomes

     ```R
     while (class(output) == "try-error" && try_number < 6) {
     ...
     }
     ```

8. It might be useful to note in the README that MonkeyLearn supports registration through GitHub, which makes the registration process really easy.

9. It might make the README a little flashier if it had some graphs. That's totally not necessary at all, but it seems like that captures peoples' attention better than pure text.

Overall, really easy to use and really nicely done.
