### 2016-07-12 #45 @noamross
One suggestion regarding outputting `tbl_df()`'s from your functions.  Per the [update](https://github.com/hadley/tibble/releases/tag/v1.1) to **tibble** and [this comment](https://twitter.com/hadleywickham/status/752830325722583040) from hadley, it may make sense to replace `tbl_df()` (as well as `as.data.frame()`) with `tibble::as_tibble()` to anticipate future changes in the tidyverse.  This won't increase the package weight as **dplyr** already depends on **tibble**. If you do so, specify `tibble (>= 1.1)` in `Imports:`.
