### 2016-07-12 #45 @masalmon
Many thanks!! 👍 I am on holidays this week so will implement the suggested changes next week. ☺
