### 2016-07-20 #45 @masalmon
Hi @leeper, thanks again, this will be very useful!

I have only questions about 2 points:

* I'm not sure I understand what you mean by "I think it would be more useful to return a single tbl_df and leave the "headers" information as attributes to that, in the form of a list"? Do you mean adding the information from request as columns to the results data frame instead of having them separately? So that there would be a column "x.query.limit.request.queries" with the same value for the whole data frame? But you say "list", so I'm pretty sure I'm misunderstanding your suggestion.

* And does the following point mean you think that the remaining number of queries should not be part of the output, or that there should be an additional warning?

I had chosen to output a list of data frames because I did it this way for [`ropenaq`](https://github.com/ropenscilabs/ropenaq) but obviously I'm open to changes. :grinning:
