### 2016-07-20 ropensci/monkeylearn#10 @maelle
add graphs to the README
See @leeper's comment in https://github.com/ropensci/onboarding/issues/45
