### 2016-07-20 #45 @noamross
@masalmon R objects have an "attributes" component that stores metadata such as dimensions and names, but also can hold arbitrary R objects as metadata. See [this section in Advanced R](http://adv-r.had.co.nz/Data-structures.html#attributes), [the R Manual section](https://cran.r-project.org/doc/manuals/r-release/R-lang.html#Attributes), `?attributes` and `?attr`.  So you can attach the headers  to your output and access it with `attributes(output)$headers` or `attr(output, "headers")`. I believe @leeper is suggesting that since this header data isn't inherently tabular and consists of a few values, just leaving it as a list makes more sense than converting it to a data frame/tibble.
