### 2016-07-20 #45 @masalmon
Okay so I now see why it (maybe) made sense to have `headers` as a tibble: when the input is a vector of texts (and not a single text), the `headers` is a table with one line per original text. It would be harder to have this if `headers` were a list. I had the same strategy in [`geoparser`](https://github.com/ropenscilabs/geoparser).

I'd tend to keep this because I don't know any better solution for this case, and if I do, I'd replace the current "text index" in the `headers` and `results` tables by a md5 checksum as I currently do in `geoparser`.

Any thought?
