### 2016-07-25 #45 @masalmon
Monkeylearn has actually started working on an API endpoint for finding public modules so I'll wait for its being ready for updating the package 😊
