### 2016-07-26 #45 @leeper
Sorry for the being incommunicado on this. What I meant in terms of headers is - as @noamross describes - that I would have the functions return something like:

```R
structure(data.frame(...), headers = list(...))
```

rather than:

```R
structure(list(results = data.frame(...), headers = data.frame(...)))
```

It seems like the headers are secondary information in most cases, so this simplifies accessing the results. Whether results are a df or tibble...either way seems fine.

Sounds good on waiting to implement a list of modules!
