### 2016-07-26 #45 @noamross
How about having the headers as a df/tibble, but still as an attribute of
the results?

On Tue, Jul 26, 2016 at 9:21 AM Maëlle Salmon <notifications@github.com>
wrote:

> No problem & thanks for your answer!
>
> But in the case where the output data.frame is actually obtained from
> binding results from several calls to the API, would you expect the headers
> to be a list of list?
>
