### 2016-07-26 #45 @masalmon
That sounds like a good compromise!
