### 2016-07-26 #45 @leeper
Yup, that works. Then there's a consistent structure regardless of whether it's one call or many.
