### 2016-07-30 #45 @masalmon
I feel I can now answer the review comments. Thanks a lot, @leeper! And thanks @noamross for your contributions to the discussion!

Something I have done independently of this feedback was to add a small vignette where I do a small analysis of the book Little Women.

1. Great point! I have added a chunk at the beginning of both vignettes now.

2. Thanks, now there is no NOTE.

3. Now the output is a data.frame with a data.frame as attributes for representing the headers.

4. In the documentation of  `monkeylearn_classify` and `monkeylearn_extract` and in the README+vignette I explain how to get the remaining number of calls, so the user could use them as they wish.

5. Thanks, I replaced `print` with `message`.

6. Thanks, this is a great suggestion. This is what I have done:

* For classifiers, when I asked again about a list of all modules, Monkeylearn was kind enough to add an endpoint, which they have not done for extractors because there are less of them. I added the function `monkeylearn_classifiers` which returns a data.frame of all classifiers, or only of private ones, with their IDs, names and other information such as the language. So the user could use this table to extract the IDs of the classifiers they want to use.

* For extractors, in the README and vignette I only mention that one can go to the website to browse existing extractors and then save their ID as an environment variable. The reason why I did not add a data.frame of some extractors myself as a .RData is that I would have to choose which characteristics of the extractors to write in this table and for some applications there are different extractors (e.g. sentiment analysis trained on Tweets vs on hotel reviews) and my list would soon be out of date (which by the way is bad for the README and vignette too)... and also I not-so-secretly hope there'll be an endpoint for getting all of them, which might happen when one can customize extractors. Is it ok if I simply wait for a future version of the API before making the extractors IDs easier to find?

7. I added the missing spaces.

8. I added the sentence "Note that MonkeyLearn supports registration through GitHub, which makes the registration process really easy.".

9. As soon as I find a really good idea, I'll add that (probably as a advertisement for a supplementary vignette about one use case)! My current idea is to maybe use the [R interface to the Urban Dictionary's API](https://github.com/Ironholds/udapi), get random definitions, and look for profanity and abuse detection, before plotting how many of the definitions have a high score. I am open to any cooler suggestion! In the meantime, is it ok I don't have any graph in the README?

Thanks again for the feedback, I'm really happy to be able to improve the package!
