### 2016-08-03 #45 @masalmon
Oh thanks a lot @noamross , fixing this.

@feconroses is it normal for headers to sometimes contain transfer.encoding and sometimes content.length?
