### 2016-08-05 #45 @masalmon
@feconroses ok thanks a lot!

@noamross has my commit solved the bug on your PC?
