### 2016-08-05 #45 @noamross
Yes, that has solved it, all checks now passing!

I hope you don't mind one more tweak at the end of a long review:.  We're starting to use Gabor Csardi's [**goodpractice**](https://github.com/MangoTheCat/goodpractice) package to check for antipatterns in code.  Running `goodpractice::gp()` yielded two small suggestions.  After these you should be good to go.

```
  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data.
    Consider using vapply() instead.

    R/utils.R:97:44
    R/utils.R:105:25
    R/utils.R:112:28

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions. They are error
    prone and result 1:0 if the expression on the right hand side is zero. Use seq_len() or seq_along()
    instead.

    R/classify.R:45:12
    R/extractor.R:70:12
```
