### 2016-08-05 #45 @masalmon
Thanks for your patience and help, @noamross, now I really got rid of `sapply`.
