### 2016-08-05 #45 @noamross
Accepted!  Thank you @masalmon and @leeper for your review and follow-up.  @masalmon, I believe you know the drill from here.
