### 2016-08-05 #45 @masalmon
@sckott could you please create the `monkeylearn` project in the ropenscilabs Appveyor account and add me as an user? thank you!
