### 2016-08-05 #45 @sckott
@masalmon the appveyor is set up
