### 2016-09-11 #45 @masalmon
For info the package is on CRAN now. Thanks again for the review!
